[runner] Split src into domain folders with a dependency-direction lint
apps/runner/src was a flat directory mixing five responsibilities
(orchestration, protocol, agent runtime, execution, filesystem). As the
per-step protocol and checkout work lands, the lack of seams was starting
to bite. Move the files into domain folders and add a biome override that
forbids the lower layers (protocol/execution/workspace/agent) from
importing orchestration, so the dependency points one way and the seam
does not erode.

No behavior change. The #* -> ./src/* import map makes the move pure
path-shuffling; only specifier strings change. StepResult is extracted
from run-step.ts into execution/step-result.ts so the shared executor
result type no longer lives inside one specific executor. The override
re-declares the base noRestrictedImports paths/patterns (biome overrides
replace rather than merge options) before appending the orchestration rule.

Refs ENG-410
