+++ apps/runner/src/execution/setup-step.ts
@@ -1,5 +1,5 @@
-import type {StepResult} from '#run-step.js';
-import {createJobDir} from '#workspace.js';
+import type {StepResult} from '#execution/step-result.js';
+import {createJobDir} from '#workspace/workspace.js';
 
 // The synthetic "Set up job" step body. It owns per-job workspace preparation and
 // reports failures through the normal step protocol, so a setup failure fails the
